feat: try no gui mode
