+++ packages/gui/extra/banner.txt
@@ -1,0 +1,8 @@
+    ____                 _____ _     __
+   / __ \___ _   __     / ___/(_)___/ /__  _________ ______
+  / / / / _ \ | / /_____\__ \/ / __  / _ \/ ___/ __ `/ ___/
+ / /_/ /  __/ |/ /_____/__/ / / /_/ /  __/ /__/ /_/ / /
+/_____/\___/|___/     /____/_/\__,_/\___/\___/\__,_/_/
+
+
+==================== 开发者边车 ====================
