+++ packages/gui/package.json
@@ -13,6 +13,7 @@
     "lint": "vue-cli-service lint",
     "electron:build": "vue-cli-service electron:build",
     "electron": "vue-cli-service electron:serve",
+    "electron:headless": "cross-env ELECTRON_RUN_AS_NODE=1 vue-cli-service electron:serve",
     "postinstall": "electron-builder install-app-deps",
     "postuninstall": "electron-builder install-app-deps",
     "electron:icons": "electron-icon-builder --input=./public/logo/win.png --output=build --flatten",
@@ -27,6 +28,7 @@
     "@vscode/sudo-prompt": "^9.3.1",
     "adm-zip": "^0.5.16",
     "ant-design-vue": "^1.7.8",
+    "cac": "^6.7.14",
     "electron-baidu-tongji": "^1.0.5",
     "electron-updater": "^6.3.9",
     "json5": "^2.2.3",
@@ -44,6 +46,7 @@
     "@vue/babel-preset-jsx": "^1.4.0",
     "@vue/cli-plugin-babel": "^5.0.8",
     "@vue/cli-service": "^5.0.8",
+    "cross-env": "^7.0.3",
     "electron": "^19.1.9",
     "electron-builder": "^25.1.8",
     "electron-icon-builder": "^2.0.1",
